Experimenting with using IDbContextFactory instead of directly injecting DbContext into EF triggers
When injecting the DbContext directly into a trigger an exception gets thrown that the service provider cannot resolve the scoped DbContext. It's hard to tell if this is a bug on our side or a bug in the library since their documentation is so piss poor. This commit aims to test whether injecting the factory instead of the context will fix the issue.

While this alleviate the issue from what I can tell, Identity does not correctly annotate the user properties for nullability meaning that the PhoneNumber is required despite not having the ability to input a phone number anywhere. This causes a database constraint failure.

I will revisit this later as having accounts is a secondary priority for the website.