Fixing the repositories to not throw when disposed
I was getting screwed by order of operations, stack traces that were conflicting with documentation, and a misunderstanding of how DbContext needs to be disposed. After a bit of messing around I believe I have finally resolved the issue. Unfortunately this means I had to tweak the repository structure a bit.

Instead of creating repositories directly on the IRepositoryProvider the IRepositoryProvider now returns a RepositoryScope which allows for committing changes to the underlying DbContext. The scope now exposes the methods for creating the repositories.

- Unsealed ThreadSafeCache and added Keys and Values accessors
- Added IAsyncDisposable support to Disposable.cs
- Added a wrapper around ActivatorUtilities.CreateInstance
- Type.IsNullable() now forwards to IsNullableCore() to correctly detect null reference types
- Fixed an issue where the default value generation on database columns were not taking into account keys and indices. Bool and numeric defaults are now added for properties that are not keys or indices and do not have defaults defined
- Implemented new repository pattern as described above
- Re-enabled TriggeredPooledDbContextFactory and triggers